rest_api_pkg/requirements.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django>=1.11.5

# Tables
django-filter>=1.0.4

# DRF
djangorestframework>=3.6.4

# Python-PostgreSQL Database Adapter
psycopg2>=2.7.3
